+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,16 +1,22 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 480
+INVENTREE_API_VERSION = 482
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
 
-v480 -> 2026-04-19 : https://github.com/inventree/InvenTree/pull/11631
+v482 -> 2026-05-03 : https://github.com/inventree/InvenTree/pull/11631
     - Adds "raw_amount" field to the BomItem API endpoint
 
+v481 -> 2026-04-28 : https://github.com/inventree/InvenTree/pull/11825
+    - Adds new "bom" ruleset and associated permissions for BOM management, separate from the "part" ruleset which remains focused on part management
+
+v480 -> 2026-04-27 : https://github.com/inventree/InvenTree/pull/11816
+    - The "issued_by" field on the Build API endpoint is now read-only, and is automatically set to the current user when a build is created
+
 v479 -> 2026-04-11 : https://github.com/inventree/InvenTree/pull/11723
-    - POST /api//notifications/readall/ now requires a POST action
+    - POST /api/notifications/readall/ now requires a POST action
     - POST /api/admin/email/test/ - now returns a 200 on. a successful test
     - GET /api/notifications/ - now uses user-centric permissions, not a general read